requirements.txt
asgiref==3.7.2
Django==4.2.6
django-crispy-forms==2.1
Pillow==10.1.0
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
